html/semantics/embedded-content-0/media-elements/loading-the-media-resource/029.html

<!doctype html>
<title>pointer updates (removing text nodes)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
var a = 0;
var b = 0;
var c = 0;
</script>
<video
 >x<source onerror=a++
 >x<source onerror=b++ src='/delay/?type=html&amp;delay=50&amp;content-type=video%2Fogg'
 >x<source onerror=c++
 >x</video
>
<script>
async_test(function(t) {
  var video = document.querySelector('video');
  // remove the text nodes
  video.removeChild(video.childNodes[0]);
  video.removeChild(video.childNodes[1]);
  video.removeChild(video.childNodes[2]);
  video.removeChild(video.childNodes[3]);
  window.onload = t.step_func(function() { 
    assert_equals(a, 1, 'error events on a');
    assert_equals(b, 1, 'error events on b');
    assert_equals(c, 1, 'error events on c');
    t.done();
  });
});
</script>
